fix: Self query issues
* Countdown tests adjusted, countdown implementation "logged in" pattern simplified.
* User self query removed (calling it by accident sync loops the app).
* User self query moved to context.
